Move Testresources to catalog-testkit
